// src/sdk/database/orm/orm_user_accounts.ts
export interface UserAccountsModel {
  id: string;
  data_creator: string;
  data_updater: string;
  create_time: string;
  update_time: string;
  email: string;
  name: string;
  is_premium: boolean;
  password: string;
}